auto/llm/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hpt-llm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "slack-bolt>=1.18",
    "python-dotenv>=1.0",
    "kafka-python>=2.0.2",
    "apscheduler>=3.10",
    "psycopg[binary]>=3.1",
    "prometheus-client>=0.20",
]

[project.scripts]
hpt-llm = "hpt_llm.bot:main"

[tool.setuptools.packages.find]
where = ["src"]
